Add files via upload
Farthest pair (in one dimension). Write a program that, given an array a[] of N
double values, finds a farthest pair: two values whose difference is no smaller than the
the difference of any other pair (in absolute value). The running time of your program
should be linear in the worst case.